+++ common/src/main/java/house/greenhouse/rapscallionsandrockhoppers/entity/behaviour/AmphibiousSetRandomSwimTarget.java
@@ -4,7 +4,7 @@ import house.greenhouse.rapscallionsandrockhoppers.entity.Penguin;
 import net.minecraft.world.entity.ai.behavior.BehaviorUtils;
 import net.minecraft.world.entity.ai.memory.MemoryModuleType;
 import net.minecraft.world.entity.ai.memory.WalkTarget;
-import net.minecraft.world.entity.ai.util.DefaultRandomPos;
+import net.minecraft.world.entity.ai.util.LandRandomPos;
 import net.minecraft.world.phys.Vec3;
 import net.tslat.smartbrainlib.api.core.behaviour.custom.path.SetRandomWalkTarget;
 import net.tslat.smartbrainlib.util.BrainUtils;
@@ -12,14 +12,10 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Predicate;
 
-public class SetRandomSwimTarget extends SetRandomWalkTarget<Penguin> {
-    protected Predicate<Penguin> avoidLandPredicate = entity -> true;
+public class AmphibiousSetRandomSwimTarget extends SetRandomWalkTarget<Penguin> {
+    protected Predicate<Penguin> moveToLandPredicate = entity -> true;
 
-    public SetRandomSwimTarget dontAvoidLand() {
-        return avoidLandWhen(entity -> false);
-    }
-
-    public SetRandomSwimTarget avoidLandWhen(Predicate<Penguin> predicate) {
+    public AmphibiousSetRandomSwimTarget moveToLandWhen(Predicate<Penguin> predicate) {
         this.avoidWaterPredicate = predicate;
         return this;
     }
@@ -34,12 +30,12 @@ public class SetRandomSwimTarget extends SetRandomWalkTarget<Penguin> {
         if (targetPos == null) {
             BrainUtils.clearMemory(penguin, MemoryModuleType.WALK_TARGET);
         } else {
-            BrainUtils.setMemory(penguin, MemoryModuleType.WALK_TARGET, new WalkTarget(targetPos, this.speedModifier.apply(penguin, targetPos), 2));
+            BrainUtils.setMemory(penguin, MemoryModuleType.WALK_TARGET, new WalkTarget(targetPos, this.speedModifier.apply(penguin, targetPos), 3));
         }
     }
 
     @Override
     protected @Nullable Vec3 getTargetPos(Penguin penguin) {
-        return this.avoidLandPredicate.test(penguin) ? BehaviorUtils.getRandomSwimmablePos(penguin, (int) this.radius.xzRadius(), (int) this.radius.yRadius()) : DefaultRandomPos.getPos(penguin, (int) this.radius.xzRadius(), (int) this.radius.yRadius());
+        return this.moveToLandPredicate.test(penguin) ? LandRandomPos.getPos(penguin, (int) this.radius.xzRadius(), (int) this.radius.yRadius()) : BehaviorUtils.getRandomSwimmablePos(penguin, (int) this.radius.xzRadius(), (int) this.radius.yRadius());
     }
 }