+++ common/src/main/java/house/greenhouse/rapscallionsandrockhoppers/entity/behaviour/PenguinJump.java
@@ -25,7 +25,7 @@ public class PenguinJump extends ExtendedBehaviour<Penguin> {
 
     @Override
     protected boolean checkExtraStartConditions(ServerLevel level, Penguin penguin) {
-        if (penguin.getTimeAllowedToWaterJump() > penguin.tickCount || penguin.getAirSupply() > 260) {
+        if (penguin.getTimeAllowedToWaterJump()  <= 0 || penguin.getAirSupply() > 260) {
             return false;
         }
 
@@ -72,7 +72,7 @@ public class PenguinJump extends ExtendedBehaviour<Penguin> {
     protected void stop(Penguin penguin) {
         this.breached = false;
         penguin.setXRot(0.0F);
-        penguin.setTimeAllowedToWaterJump(Optional.of(penguin.tickCount + Mth.randomBetweenInclusive(penguin.getRandom(), 400, 600)));
+        penguin.setTimeAllowedToWaterJump(Optional.of(Mth.randomBetweenInclusive(penguin.getRandom(), 400, 600)));
         BrainUtils.clearMemories(penguin, RockhoppersMemoryModuleTypes.IS_JUMPING);
     }