+++ common/src/main/java/house/greenhouse/rapscallionsandrockhoppers/entity/behaviour/PenguinPeck.java
@@ -36,8 +36,8 @@ public class PenguinPeck extends AnimatableMeleeAttack<Penguin> {
                 penguin.doHurtTarget(this.target);
                 if (!this.target.isAlive()) {
                     this.target.remove(Entity.RemovalReason.KILLED);
-                    penguin.setHungryTime(Optional.of(penguin.tickCount + 4800));
-                    penguin.setTimeAllowedToEat(Optional.of(penguin.tickCount + 400));
+                    penguin.setHungryTime(Optional.of(4800));
+                    penguin.setTimeAllowedToEat(Optional.of(400));
                     penguin.incrementFishEaten();
                     BrainUtils.clearMemory(penguin, MemoryModuleType.ATTACK_TARGET);
                     BrainUtils.clearMemory(penguin, MemoryModuleType.WALK_TARGET);