+++ common/src/main/java/house/greenhouse/rapscallionsandrockhoppers/entity/behaviour/WalkToRewardedPlayer.java
@@ -49,7 +49,7 @@ public class WalkToRewardedPlayer extends ExtendedBehaviour<Penguin> {
 
     @Override
     protected void start(Penguin penguin) {
-        BrainUtils.setMemory(penguin, MemoryModuleType.WALK_TARGET, new WalkTarget(this.runPos, 1.0F, 0));
+        BrainUtils.setMemory(penguin, MemoryModuleType.WALK_TARGET, new WalkTarget(this.runPos, 1.0F, 1));
     }
 
     @Override