+++ common/src/main/java/house/greenhouse/rapscallionsandrockhoppers/mixin/BoatMixin.java
@@ -44,7 +44,7 @@ public abstract class BoatMixin extends VehicleEntity {
     @Inject(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/vehicle/Boat;controlBoat()V", shift = At.Shift.BY, by = 2))
     private void rapscallionsandrockhoppers$addPenguinSpeedBonus(CallbackInfo ci) {
         BoatPenguinsAttachment boatData = RapscallionsAndRockhoppers.getHelper().getBoatPenguinData((Boat)(Object)this);
-        if (this.getStatus().equals(Boat.Status.IN_WATER) && boatData.penguinCount() > 0 && boatData.getFollowingPenguins().stream().anyMatch(uuid -> {
+        if (boatData.penguinCount() > 0 && boatData.getFollowingPenguins().stream().anyMatch(uuid -> {
             Entity entity = EntityGetUtil.getEntityFromUuid(this.level(), uuid);
             return entity != null && entity.isInWater();
         })) {